feat(filters): removed Category page from uikit
